langchain_search/requirements.txt
streamlit
langchain
google-search-results

langchain-google-genai